+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * fix parsing newline character literal
 * fix reading local files in browser [#276](https://github.com/jcubic/lips/issues/276)
 * fix parsing invalid expression `(1 . 2 3)` [#245](https://github.com/jcubic/lips/issues/245)
+* fix invalid error message for not matched `syntax-rules` [#243](https://github.com/jcubic/lips/issues/243)
 
 ## 1.0.0-beta.17
 ### Breaking